refactor: extract CoursesContext
