+++ src/contexts/CoursesContext.js
@@ -1,0 +1,7 @@
+import { createContext } from 'react';
+
+export const CoursesContext = createContext({
+	data: null,
+	loading: false,
+	error: null,
+});
